Fetch blocks concurrently but send in order (#32)
This PR rearchitects the main loop such that we can run multiple
goroutines to fetch blocks concurrently, but send them in order to the
API to ensure we don't get gaps. We produce block numbers to fetch on an
unbuffered channel (`ProduceBlockNumbers`), and each concurrent
`ConsumeBlock` goroutine gets a block number from that channel. The
`SendBlocks` goroutine receives all blocks on an unbuffered channel, but
buffers them in a map until they can be sent in order.

I deleted the test that asserted the exact number of blocks that were
sent. I couldn't manage to get it working without off by one errors
sometimes. Since the main use case is running it forever and not with a
given blocks, I figured this was an OK decision. Let me know.

Possibly problematic scenario: If block N - and only block N - is
somehow problematic from the RPC node, and must be retried ~forever,
what will happen? MaxBatchSize - 1 goroutines continue to fetch blocks
from the RPC node successfully, causing the buffer in SendBlocks to fill
up. The mechanism in this PR doesn't have a way to backpressure there.
Is that a problem? If memory footprint is a concern, we could possibly
compress it here: At this point, currently, the block response is
uncompressed. If we compress it first, it would take a lot more blocks
for this to consume a lot of memory.

Otherwise, if there are general problems with the RPC node such that all
requests take a long time or need to be retried, we'll get backpressured
because we limit the concurrent number of calls to the node to
MaxBatchSize.

FWIW I'm running this locally and getting 12 blocks per second with 5
workers!