Send batch of blocks from the main loop (#37)
This PR changes the node indexer from sending one block at a time to
sending a batch of blocks. Earlier we implemented concurrent block
fetching with buffering (#32). On a configurable interval (defaults to
every second), we now check the buffer and send all possible blocks.

We add a flag for specifying whether you wish to skip a failed block or
not. It's off by default. This means if all the retries to the RPC node
fails for a given block, we will crash. This ensures no block gaps.

---------

Co-authored-by: Miguel Filipe <miguel.filipe@gmail.com>