Add error reporting to progress report API call (#44)
This PR makes it so we collect errors and send them to Dune as part of
the progress report request.

I added two models in this PR, one in the `models` package and one in
the DuneAPI package. This means we have three structs for representing
the same kind of error (there's also one in `ingester). Is that too
much? 😅