Use a constant amount of memory for error slices and add error abstraction (#57)
This PR changes the Info struct in the Ingester a bit so that we have an
inner Errors struct for keeping track of errors. We add a method for
resetting at, for observing errors, and for transforming to progress
reports. We make sure that the error slices do not grow unbounded, but
we keep track of the total error counts as well.