DuneAPI: new batch wire format (breaking change!)
This is a breaking change, it changes the block batch format on the wire.
It introduces a single JSON line (NDJSON) which is the batch header.
This can be DISABLED with `DUNEAPI_DISABLE_BATCH_HEADER`
environment variable

This line indicates how many blocks and the size of each block message
In essense, we're length prefixing the batch.
This allows the reader to split out easily all the messages.
Until now this wasn't needed because Optimism Stack (EVM stack) uses a
fixed 3-line payload for the blocks.
But this isn't the case in other EVM stacks (such as arbitrum nitro)

It uses JSON to encode this information because the jsonRPC message from
the blockchain RPC node are already in JSON and the wire format is already NDJSON.

For the server side to handle in a backwards compatible way the payload,
it needs to read the first line and attempt to parse it as a BlockBatchHeader message
If that fails, it can revert back to the older message format parser