refactor: implement a maximum upperbound on parallel http requests to the RPC node
Currently we allow defining RPCConcurrency, but that is in reality the
BLOCK CONCURRENCY, the number of blocks we fetch in parallel.

for OP-Stack EVMs, we issue, for each block, in parallel, 3 RPC requests
to the RPC Node.
So, the real concurrency is something like:
3 * (RPCConcurrency + DLQRPCConcurrency)

This patch enforces a maximum concurrency by using on rpcClient an
internal workerPool.
This uses the github.com/panjf2000/ants/v2 library

Now the maximum total RPC concurrency is defined as 4 x RPCConcurrency